requirements.txt
Flask>=2.0
flask-cors
python-dotenv
pdfplumber
pytesseract
pdf2image
google-generativeai

# pip install -r requirements.txt
